+++ felupe/__init__.py
@@ -103,6 +103,8 @@ from .tools import (
 )
 from .mechanics import (
     SolidBody,
+    SolidBodyNearlyIncompressible,
+    StateNearlyIncompressible,
     SolidBodyPressure,
     SolidBodyTensor,
     SolidBodyGravity,