+++ felupe/mechanics/__init__.py
@@ -1,4 +1,8 @@
 from ._solidbody import SolidBody
+from ._solidbody_incompressible import (
+    SolidBodyNearlyIncompressible,
+    StateNearlyIncompressible,
+)
 from ._solidbody_pressure import SolidBodyPressure
 from ._solidbody_tensor import SolidBodyTensor
 from ._solidbody_gravity import SolidBodyGravity