+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [Unreleased]
 
+## [5.3.0] - 2022-11-03
+
 ### Added
 - Add optional pre-compression to shear-loadcase `dof.shear(compression=0.0)`.
 - Add `MeshContainer` and string-representation for `Mesh` objects.